Document dark forms (#37)
* Upgrade gosheets module to v0.9.1

* Implement dark forms in off canvas

* Adds documentation for dark forms

Dark forms are mostly like the default forms with a few alterations
based on the state of the input. Labels and text show up as a light
color so they are legible on a dark background. When the input is
in a disabled state, the input will darken slightly to better
portray itself as such.